requirements.txt
torch==0.3.1
torchvision==0.2.1
visdom==0.1.8.9
numpy==1.14.2
h5py==2.7.1
matplotlib==2.1.2
Pillow==5.0.0
scikit-image==0.14.0
scipy==1.3.2
pandas
